+++ src/main/java/HibernateSessionFactoryUtil.java
@@ -1,26 +1,23 @@
-import hibernate.model.Employee;
-import org.hibernate.SessionFactory;
-import org.hibernate.boot.registry.StandardServiceRegistryBuilder;
-import org.hibernate.cfg.Configuration;
-
-
-public class HibernateSessionFactoryUtil {
-    private static SessionFactory sessionFactory;
-    private HibernateSessionFactoryUtil() {}
-
-    public static SessionFactory getSessionFactory() {
-        if (sessionFactory == null) {
-            try {
-                Configuration configuration = new Configuration().configure();
-                configuration.addAnnotatedClass(Employee.class);
-                StandardServiceRegistryBuilder builder = new StandardServiceRegistryBuilder()
-                        .applySettings(configuration.getProperties());
-                sessionFactory = configuration.buildSessionFactory(builder.build());
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-        }
-        return sessionFactory;
-    }
-
-}
+//import org.hibernate.boot.registry.StandardServiceRegistryBuilder;
+//import org.hibernate.cfg.Configuration;
+//
+//
+//public class HibernateSessionFactoryUtil {
+//    private static SessionFactory sessionFactory;
+//    private HibernateSessionFactoryUtil() {}
+//
+//    public static SessionFactory getSessionFactory() {
+//        if (sessionFactory == null) {
+//            try {
+//                Configuration configuration = new Configuration().configure();
+//                configuration.addAnnotatedClass(Employee.class);
+//                StandardServiceRegistryBuilder builder = new StandardServiceRegistryBuilder()
+//                        .applySettings(configuration.getProperties());
+//                sessionFactory = configuration.buildSessionFactory(builder.build());
+//            } catch (Exception e) {
+//                e.printStackTrace();
+//            }
+//        }
+//        return sessionFactory;
+//    }
+//}
