+++ src/main/java/service/EmployeeDao.java
@@ -5,13 +5,14 @@ import hibernate.model.Employee;
 import java.util.List;
 
 public interface EmployeeDao {
+    // добавление сущности
     Integer add(Employee employee);
-
+    // получение по id
     Employee getById(int id);
-
+    // получение всех сотрудников
     List<Employee> getAllEmployee();
-
+    // добавление сотрудника
     void updateEmployee(Employee employee, int id);
-
+    // удаление сотрудника
     void deleteEmployee(Employee employee);
 }
