+++ src/main/java/service/EmployeeDaoImpl.java
@@ -3,7 +3,6 @@ package service;
 import hibernate.model.Employee;
 import org.hibernate.Session;
 import org.hibernate.Transaction;
-import service.EmployeeDao;
 
 import java.util.List;
